requirements.txt
django-celery-beat
celery
redis
django-celery-results
